changed datatype from <f8 to float for hdf5 images

+++ dataset.py
@@ -99,7 +99,7 @@ def pil_and_hdf5_loader(path):
     # for hdf5
     if(path.endswith('.hdf5')):
         with h5py.File(path, 'r') as f:
-            return f[HDF5_DATASET_NAME][:]
+            return f[HDF5_DATASET_NAME][:].astype(float)
 
             # note:
             # DONOT USE: np.array(f[hdf5_dataset_name]) it was much slower in testi
